skills/song-production/scripts/pyproject.toml
[project]
name = "music-project"
version = "0.1.0"
description = "火山引擎音乐生成项目"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "imageio-ffmpeg>=0.5.0",
]
